Updates bundles reducer
