+++ src/state/actions/index.ts
@@ -45,7 +45,7 @@ export interface BundleCompleteAction {
         cellId: string;
         bundle: {
             code: string;
-            err: string;
+            error: string;
         }
     }
 }
