+++ src/state/reducers/bundlesReducer.ts
@@ -6,7 +6,7 @@ interface BundlesState {
     [key: string]: {
         loading: boolean;
         code: string;
-        err: string;
+        error: string;
     } | undefined;
 }
 
@@ -18,7 +18,7 @@ const reducer = produce((state: BundlesState = initialState, action: Action): Bu
             state[action.payload.cellId] = {
                 loading: true,
                 code: '',
-                err: '',
+                error: '',
             
             };
             return state;
@@ -26,7 +26,7 @@ const reducer = produce((state: BundlesState = initialState, action: Action): Bu
             state[action.payload.cellId] = {
                 loading: false,
                 code: action.payload.bundle.code,
-                err: action.payload.bundle.err,
+                error: action.payload.bundle.error,
             };
             return state;
         default:
